Merge dev into master (frontend refactoring)

+++ GiEnJul/Controllers/AdminController.cs
@@ -8,6 +8,7 @@ using GiEnJul.Infrastructure;
 using GiEnJul.Models;
 using GiEnJul.Repositories;
 using GiEnJul.Utilities;
+using GiEnJul.Utilities.EmailTemplates;
 using GiEnJul.Utilities.ExcelClasses;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -32,6 +33,7 @@ namespace GiEnJul.Controllers
         private readonly IMapper _mapper;
         private readonly IEmailClient _emailClient;
         private readonly ISettings _settings;
+        private readonly IEmailTemplateBuilder _emailTemplateBuilder;
 
         public AdminController(
             IEventRepository eventRepository,
@@ -42,7 +44,8 @@ namespace GiEnJul.Controllers
             ILogger log,
             IMapper mapper,
             IEmailClient emailClient,
-            ISettings settings)
+            ISettings settings,
+            IEmailTemplateBuilder emailTemplateBuilder)
         {
             _eventRepository = eventRepository;
             _giverRepository = giverRepository;
@@ -53,6 +56,7 @@ namespace GiEnJul.Controllers
             _mapper = mapper;
             _emailClient = emailClient;
             _settings = settings;
+            _emailTemplateBuilder = emailTemplateBuilder;
         }
 
         [HttpGet("Overview/Givers")]
@@ -61,6 +65,15 @@ namespace GiEnJul.Controllers
         {
             var activeEvent = await _eventRepository.GetActiveEventForLocationAsync(location);
             var givers = await _giverRepository.GetGiversByLocationAsync(activeEvent, location);
+
+            var recipients = await _recipientRepository.GetRecipientsByLocationAsync(activeEvent, location);
+
+            foreach (var giver in givers.Where(x => x.IsSuggestedMatch))
+            {
+                var matchedRecipient = recipients.Find(x => x.RowKey == giver.MatchedRecipient);
+                giver.MatchedFamilyId = matchedRecipient.FamilyId;
+            }
+
             return givers
                 .OrderBy(x => x.HasConfirmedMatch)
                 .ThenBy(x => x.IsSuggestedMatch)
@@ -133,18 +146,28 @@ namespace GiEnJul.Controllers
 
         [HttpDelete("Connection")]
         [Authorize(Policy = "DeleteConnection")]
-        public async Task<ActionResult> DeleteConnectionAsync(string location, string rowKey)
+        public async Task<ActionResult> DeleteConnectionAsync([FromBody] DeleteConnectionDto connectionDto)
         {
-            var giver = await _giverRepository.GetGiverAsync(location, rowKey);
+            var giver = await _giverRepository.GetGiverAsync(connectionDto.PartitionKey, connectionDto.RowKey);
 
-            if (giver?.MatchedRecipient is null)
+            var recipient = await _recipientRepository.GetRecipientAsync(connectionDto.PartitionKey, connectionDto.RowKey);
+
+            if (giver is null && recipient is null)
             {
                 return NotFound();
             }
 
-            var recipient = await _recipientRepository.GetRecipientAsync(location, giver.MatchedRecipient);
+            if (giver?.MatchedRecipient != null && recipient is null)
+            {
+                recipient = await _recipientRepository.GetRecipientAsync(connectionDto.PartitionKey, giver.MatchedRecipient);
+            }
 
-            if (recipient is null)
+            if (recipient?.MatchedGiver != null && giver is null)
+            {
+                giver = await _giverRepository.GetGiverAsync(connectionDto.PartitionKey, recipient.MatchedGiver);
+            }
+
+            if (recipient is null || giver is null)
             {
                 return NotFound();
             }
@@ -164,8 +187,6 @@ namespace GiEnJul.Controllers
             {
                 await _giverRepository.InsertOrReplaceAsync(giver);
                 await _recipientRepository.InsertOrReplaceAsync(recipient);
-                
-                await _connectionRepository.DeleteConnectionAsync(location, recipient.RowKey + "_" + giver.RowKey);
             }
             catch (Exception e)
             {
@@ -175,7 +196,22 @@ namespace GiEnJul.Controllers
                 _log.Error(e, "Could not delete connection between {@0} and {@1}", giver, recipient);
                 return NotFound();
             }
+
+            if (!originalGiver.HasConfirmedMatch) return Ok();
             
+            try
+            {
+                await _connectionRepository.DeleteConnectionAsync(connectionDto.PartitionKey, recipient.RowKey + "_" + giver.RowKey);
+            }
+            catch (Exception e)
+            {
+                await _giverRepository.InsertOrReplaceAsync(originalGiver);
+                await _recipientRepository.InsertOrReplaceAsync(originalRecipient);
+
+                _log.Error(e, "Could not delete connection between {@0} and {@1}", giver, recipient);
+                return NotFound();
+            }
+
             return Ok();
         }
 
@@ -193,10 +229,11 @@ namespace GiEnJul.Controllers
 
             try
             {
-                var eventDto = await _eventRepository.GetEventByUserLocationAsync(giver.Location);
+                var @event = await _eventRepository.GetEventByUserLocationAsync(giver.Location);
 
                 giver.IsSuggestedMatch = true;
                 giver.MatchedRecipient = connectionDto.RecipientRowKey;
+                giver.MatchedFamilyId = recipient.FamilyId;
                 await _giverRepository.InsertOrReplaceAsync(giver);
 
                 recipient.IsSuggestedMatch = true;
@@ -205,7 +242,6 @@ namespace GiEnJul.Controllers
 
                 recipient.FamilyMembers = await _personRepository.GetAllByRecipientId(recipient.RowKey);
 
-                var title = "Gi en jul-familie - husk å bekrefte!";
                 var verifyLink = $"{_settings.ReactAppUri}/{giver.RowKey}/{recipient.RowKey}/{giver.PartitionKey}";
 
                 var recipientNote = string.IsNullOrWhiteSpace(recipient.Note) ? "" : $"<strong>Merk:</strong> {recipient.Note}";
@@ -220,55 +256,27 @@ namespace GiEnJul.Controllers
                         familyTable += " ";
                     }
                 }
+                
+                var emailTemplatename = EmailTemplateName.AssignedFamily;
+                var emailValuesDict = new Dictionary<string, string> 
+                { 
+                    { "familyTable", familyTable }, 
+                    { "verifyLink", verifyLink },
+                    { "recipientNote", recipientNote },
+                };
+                emailValuesDict.AddDictionary(ObjectToDictionaryHelper.MakeStringValueDict(giver, "giver."));
+                emailValuesDict.AddDictionary(ObjectToDictionaryHelper.MakeStringValueDict(@event, "eventDto."));
+                emailValuesDict.AddDictionary(ObjectToDictionaryHelper.MakeStringValueDict(recipient, "recipient."));
 
-                var body =
-                    $"Hei, {giver.FullName} <br/><br/> " +
+                var emailTemplate = await _emailTemplateBuilder.GetEmailTemplate(emailTemplatename, emailValuesDict);
 
-                    $"Da har vi en familie til deg! Når du har lest gjennom teksten er det viktig at du klikker på <a href='{verifyLink}'> denne linken </a> for å bekrefte at du gir familien en jul. " +
-                    $"Din familie har nummer {recipient.FamilyId}. Dette nummeret må du skrive godt synlig på esken. Ikke pakk inn eller levér noe i plastposer, men i esker som er enkle å bære. <br/><br/>" +
-
-                    $" <h3>OVERSIKT OVER FAMILIE OG GAVEØNSKER </h3>" +
-                    
-                    $"<ul>{familyTable}</ul>" +
-
-                    $"<strong>Middag:</strong>  {recipient.Dinner}<br/>" +
-                    $"<strong>Dessert:</strong> {recipient.Dessert}<br/>" +
-                    $"{recipientNote} <br/><br/> " +
-
-                    $"Juleeskene skal i år leveres {eventDto.DeliveryTime}, {eventDto.DeliveryDate} til {eventDto.DeliveryAddress}. <br/><br/>" +
-
-                    $"Juleeskene skal minst inneholde en julemiddag med dessert og en gave til hver av familiemedlemmene. Dersom du ønsker, kan du bidra med én ekstra middag og/eller noe til julefrokosten. " +
-                    $"Ekstra-middagen kan eksempelvis være pølse og potetmos, medisterkaker og poteter, kjøttdeig og spagetti, eller noe annet du synes er passende. <br/><br/>" +
-
-                    $"Har du lyst til å legge mer oppi esken, er det selvsagt frivillig. Tips: <br/><br/>" +
-                    $"<ul> <li> julestrømpe med godteri til barna </li><li> saft, juice, melk, te, kaffe </li>" +
-                    $"<li> frukt </li><li> snacks og julegodteri</li><li> julekaker</li><li> pålegg: Nugatti, leverpostei, kjøttpålegg, ost og så videre.</li>" +
-                    $"<li> servietter, lys og julepynt</li><li> brød, julekake</li></ul><br/>" +
-                    $"Pass på at ikke maten blir dårlig/sur, og vær obs på datostempel. Ikke kjøp alkoholholdig drikke! " +
-                    $"Merk også at dersom familien spiser halal, betyr det at de ikke spiser svin- og da heller ikke pålegg eller godteri og annet som inneholder svin og gelatin. <br/><br/>" +
-
-                    $"Gaver pakkes inn og merkes med til mor, til far, til jente x år, til gutt x år og så videre. " +
-                    $"Det er lurt å legge byttelapp oppi. Pakk gjerne i bananesker, eller andre esker som er lette å bære. <br/><br/>" +
-
-                    $"NB! Dersom du ønsker å gi bort brukte leker eller tøy, er det viktig at dette er i god stand, og ikke erstatter julegaven. " +
-                    $"Vi støtter selvsagt gjenbruk, men dette er familier som sjeldent kan unne seg nye ting. Dersom du har mye klær og/eller leker du ønsker å gi, men ikke matcher familien du har fått, "+ 
-                    $"kan du sende en mail til oss, så kan vi se om vi kan få gitt det til en passende familie.<br/><br/>" +
-
-                    $"Igjen tusen takk for at du er med på årets Gi en jul! Husk å følge med på <a href={eventDto.Facebook}>Facebook-eventet</a> hvor det kommer oppdateringer. <br/><br/>" +
-
-                    $"<b>Viktig</b>: Klikk på <a href='{verifyLink}'> denne linken </a> for å bekrefte at du gir familien en jul. <br/><br/>" +
-
-                    $"<b>PS</b>: Denne mailen kan ikke besvares. Ved spørsmål angående registreringen eller lignende, ta kontakt med {eventDto.ContactPerson} på <a href=\"mailto:{eventDto.Email}\">{eventDto.Email}</a> <br/><br/>" +
-
-                    $"Hilsen {eventDto.ContactPerson} i Gi en jul {eventDto.City}";
-
-
-                await _emailClient.SendEmailAsync(giver.Email, giver.FullName, title, body);
+                await _emailClient.SendEmailAsync(giver.Email, giver.FullName, emailTemplate.Subject, emailTemplate.Content);
             }
             catch (Exception e)
             {
                 giver.IsSuggestedMatch = false;
                 giver.MatchedRecipient = "";
+                giver.MatchedFamilyId = "";
                 await _giverRepository.InsertOrReplaceAsync(giver);
 
                 recipient.IsSuggestedMatch = false;
@@ -281,16 +289,47 @@ namespace GiEnJul.Controllers
             return Ok();
         }
 
-        [HttpDelete("recipient")]
+        [HttpDelete("Recipient")]
         [Authorize(Policy = "DeleteRecipient")]
         public async Task<ActionResult> DeleteRecipientAsync([FromBody] DeleteRecipientDto recipientDto)
         {
             var recipientToDelete = await _recipientRepository.GetRecipientAsync(recipientDto.PartitionKey, recipientDto.RowKey);
-            var deletedRecipient = await _recipientRepository.DeleteAsync(recipientToDelete);
+
+            if (recipientToDelete.IsSuggestedMatch)
+            {
+                await DeleteConnectionAsync(new DeleteConnectionDto(recipientDto.PartitionKey, recipientDto.RowKey));
+            }
+
+            var personsToDelete = await _personRepository.GetAllByRecipientId(recipientDto.RowKey);
+
+            if (recipientToDelete.PersonCount == 0 || personsToDelete.Count() == 0)
+            {
+                return NotFound();
+            }
+
+            await _recipientRepository.DeleteAsync(recipientToDelete);
+
+            var deleteCount = 0;
+            try
+            {
+                foreach (var person in personsToDelete)
+                {
+                    await _personRepository.DeleteAsync(person);
+                    deleteCount += 1;
+                }
+            }
+            catch (Exception e)
+            {
+                await _personRepository.InsertOrReplaceBatchAsync(personsToDelete.GetRange(0, deleteCount));
+                await _recipientRepository.InsertOrReplaceAsync(recipientToDelete);
+
+                throw e;
+            }
+
             return Ok();
         }
 
-        [HttpPut("recipient")]
+        [HttpPut("Recipient")]
         [Authorize(Policy = "UpdateRecipient")]
         public async Task<ActionResult> PutRecipientAsync([FromBody] PutRecipientDto recipientDto)
         {
@@ -319,12 +358,19 @@ namespace GiEnJul.Controllers
             return Ok();
         }
 
-        [HttpDelete("giver")]
+        [HttpDelete("Giver")]
         [Authorize(Policy = "DeleteGiver")]
         public async Task<ActionResult> DeleteGiverAsync([FromBody] DeleteGiverDto giverDto)
         {
+            var giver = await _giverRepository.GetGiverAsync(giverDto.PartitionKey, giverDto.RowKey);
+
+            if (giver.IsSuggestedMatch)
+            {
+                await DeleteConnectionAsync(new DeleteConnectionDto(giverDto.PartitionKey, giverDto.RowKey));
+            }
+
             var giverToDelete = await _giverRepository.GetGiverAsync(giverDto.PartitionKey, giverDto.RowKey);
-            var deletedGiver = await _giverRepository.DeleteAsync(giverToDelete);
+            await _giverRepository.DeleteAsync(giverToDelete);
             return Ok();
         }
